+++ app/views/__init__.py
@@ -20,6 +20,7 @@ from flask_login import current_user
 
 from app.extensions import db
 from app.views import (
+    admin,
     auth,
     database,
     order,
@@ -70,6 +71,7 @@ def init_app(app: Flask) -> None:
     app.register_blueprint(post.blueprint)
     app.register_blueprint(redirect.blueprint)
     app.register_blueprint(order.blueprint)
+    app.register_blueprint(admin.blueprint)
     app.add_url_rule("/", endpoint="index")
     database.init_app(app)
     app.after_request(format_html)  # type: ignore