+++ app/views/admin.py
@@ -1,0 +1,16 @@
+"""
+app.views.admin
+===============
+"""
+from flask import Blueprint, render_template
+
+blueprint = Blueprint("admin", __name__, url_prefix="/admin")
+
+
+@blueprint.route("/")
+def admin() -> str:
+    """Render admin page.
+
+    :return: Rendered admin page.
+    """
+    return render_template("admin.html")